reviewer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Nour Ben Gaied <157380845+NBGaied-Firemind@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
mangum = "^0.19.0"
beautifulsoup4 = "^4.13.3"
atlassian-python-api = "^3.41.19"
fastapi = {version = "^0.115.8", platform = "linux"}
boto3 = "^1.37.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
